[WINDOWS/FRAME] Added frame functionality (VariableNUmber) and window… (#1030)
* [WINDOWS/FRAME] Added frame functionality (VariableNUmber) and windows fixes.

* [IIterator] Allow for the container to be set after creation.

+++ Source/messaging/TextMessage.h
@@ -47,7 +47,7 @@ namespace Messaging {
 
             } else {
                 writer.NullTerminatedText(_text);
-                length = static_cast<uint32_t>(_text.size() + 1);
+                length = static_cast<uint16_t>(_text.size() + 1);
             }
 
             return length;
@@ -59,7 +59,7 @@ namespace Messaging {
 
             _text = reader.NullTerminatedText();
 
-            return (static_cast<uint32_t>(_text.size() + 1));
+            return (static_cast<uint16_t>(_text.size() + 1));
         }
 
         void ToString(string& text) const override
